requirements.txt
discord.py
PyNaCl==1.4.0
pandas
dnspython==2.1.0
async-timeout==3.0.1
asyncpg==0.22.0
Pillow
motor
pytz